requirements.txt
pytest
selenium
allure-pytest
delayed-assert
pytest-rerunfailures
pytest-variables
pytest-xdist
assertlib
